third_party/WebKit/LayoutTests/platform/mac-retina/virtual/mojo-loading/http/tests/security/media-element-audio-source-node-cross-origin-allowed-expected.txt

CONSOLE ERROR: line 34: Uncaught (in promise) InvalidStateError: cannot startRendering when an OfflineAudioContext is running
MediaElementAudioSourceNode with cross-origin source

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Correctly found some non-zero values.
PASS successfullyParsed is true

TEST COMPLETE
